Initiate your project folder

This function creates your project folder.

Usage

initiate_project(
  path,
  project = "single_study",
  preregistration = "empty",
  dependencies = "groundhog",
  create_github = TRUE,
  private_repo = TRUE,
  ...
)

Arguments

path

Character, the directory in which the project is initiated.

project

Character, the type of project that will be initiated. Options include 'single_study' for a project consisting of a single study, 'multistudy' for a project consisting of multiple studies, and 'registered_report' for a project containing a Registered Report. Default: 'single_study'.

preregistration

Character, The preregistration template that you want to use.

dependencies

Character, indicates whether and how you want to keep track of package dependencies. Options include "renv", "groundhog", or "none". Default: 'groundhog'

create_github

Character, should a remote repository be created on GitHub? If 'yes', a remote repository is created with the same name as the local project, and all current changes are pushed to this remote repository. If 'no', git is only set up locally. Default: 'yes'

private_repo

Logical, Should the created GitHub repository be private or public? Default: 'yes'

...

Additional arguments passed to and from functions.

Value

The initial set-up of your project repository.